Bug 1184272: require well-formed DNF scopesets
It's difficult for humans to distinguish [scope1, scope2] from [[scope1,
scope2]] during a review, but the two mean very different things.  To
avoid the risk of missing such distinctions on review, scopeMatch now
rejects the latter, or any other "shorthand" that is not a full array of
arrays of strings.